+++ auth/policies.go
@@ -155,7 +155,7 @@ type Authz interface {
 	ListAllObjects(ctx context.Context, pr PolicyReq) (PolicyPage, error)
 
 	// CountPolicies count policies based on the given PolicyReq structure.
-	CountObjects(ctx context.Context, pr PolicyReq) (int, error)
+	CountObjects(ctx context.Context, pr PolicyReq) (uint64, error)
 
 	// ListSubjects lists subjects based on the given PolicyReq structure.
 	ListSubjects(ctx context.Context, pr PolicyReq, nextPageToken string, limit uint64) (PolicyPage, error)
@@ -164,7 +164,7 @@ type Authz interface {
 	ListAllSubjects(ctx context.Context, pr PolicyReq) (PolicyPage, error)
 
 	// CountSubjects count policies based on the given PolicyReq structure.
-	CountSubjects(ctx context.Context, pr PolicyReq) (int, error)
+	CountSubjects(ctx context.Context, pr PolicyReq) (uint64, error)
 
 	// ListPermissions lists permission betweeen given subject and object .
 	ListPermissions(ctx context.Context, pr PolicyReq, filterPermission []string) (Permissions, error)
@@ -202,7 +202,7 @@ type PolicyAgent interface {
 	RetrieveAllObjects(ctx context.Context, pr PolicyReq) ([]PolicyRes, error)
 
 	// RetrieveAllObjectsCount
-	RetrieveAllObjectsCount(ctx context.Context, pr PolicyReq) (int, error)
+	RetrieveAllObjectsCount(ctx context.Context, pr PolicyReq) (uint64, error)
 
 	// RetrieveSubjects
 	RetrieveSubjects(ctx context.Context, pr PolicyReq, nextPageToken string, limit uint64) ([]PolicyRes, string, error)
@@ -211,7 +211,7 @@ type PolicyAgent interface {
 	RetrieveAllSubjects(ctx context.Context, pr PolicyReq) ([]PolicyRes, error)
 
 	// RetrieveAllSubjectsCount
-	RetrieveAllSubjectsCount(ctx context.Context, pr PolicyReq) (int, error)
+	RetrieveAllSubjectsCount(ctx context.Context, pr PolicyReq) (uint64, error)
 
 	// (ctx context.Context, pr PolicyReq, filterPermissions []string) ([]PolicyReq, error)
 	RetrievePermissions(ctx context.Context, pr PolicyReq, filterPermission []string) (Permissions, error)
